Extends webhook with stats
