+++ package.json
@@ -25,12 +25,14 @@
     "@adminjs/express": "^6.1.1",
     "@adminjs/nestjs": "6.1.0",
     "@adminjs/typeorm": "^5.0.1",
+    "@faker-js/faker": "^9.8.0",
     "@nestjs/axios": "^3.1.3",
     "@nestjs/cache-manager": "^3.0.1",
     "@nestjs/common": "^10.4.17",
     "@nestjs/config": "^4.0.2",
     "@nestjs/core": "^10.4.17",
     "@nestjs/platform-express": "^10.4.17",
+    "@nestjs/schedule": "^6.0.0",
     "@nestjs/swagger": "^8.1.1",
     "@nestjs/terminus": "^10.3.0",
     "@nestjs/typeorm": "^10.0.2",
